+++ src/enums.ts
@@ -342,7 +342,7 @@ export function isValueInStringEnum<E extends string>(strEnum: Record<string, E>
 /** True if argument is a member of CacheKey. */
 export const isCacheKey = isValueInStringEnum(CacheKey);
 /** True if argument is a member of TagTootsCacheKey. */
-export const isTagTootsCacheKey = isValueInStringEnum(TagTootsCategory);
+export const isTagTootsCategory = isValueInStringEnum(TagTootsCategory);
 
 /** True if argument is a member of NonScoreWeightName enum. */
 export const isNonScoreWeightName = isValueInStringEnum(NonScoreWeightName);